product-service/src/main/java/com/example/product/endpoint/ProductResource.java
package com.example.product.endpoint;

import com.example.product.model.ProductDTO;
import com.example.product.service.ProductService;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.server.ResponseStatusException;

import java.util.List;

@RestController
@RequestMapping("/products")
public class ProductResource {
    private final ProductService productService;

    public ProductResource(ProductService productService) {
        this.productService = productService;
    }

    @GetMapping
    List<ProductDTO> getAllProducts() {
        return productService.getAllProducts();
    }

    @GetMapping("/{id}")
    public ProductDTO getProductById (@PathVariable long id) {
        return productService.getProductById(id).orElseThrow(()-> new ResponseStatusException(
                HttpStatus.NOT_FOUND, "Product Not Found"));
    }
}
